Создал новый бранч, для части 2. Создал интерфейс для таск менеджера
